moving backup config functions to service package
